Refactor Unicode utilities to unicode_toolkit
