+++ mappings_endpoint.py
@@ -4,7 +4,7 @@ from config.service_registration import register_upload_services
 
 # Shared container ensures services are available across blueprints
 from core.container import container
-from core.unicode_handler import clean_unicode_surrogates
+from unicode_toolkit import clean_unicode_surrogates
 
 if not container.has("upload_processor"):
     register_upload_services(container)
