+++ services/upload/__init__.py
@@ -18,7 +18,7 @@ from .core.processor import UploadProcessingService
 from .core.validator import ClientSideValidator as UploadValidator
 from utils.upload_store import UploadedDataStore as UploadStorage
 from .controllers.upload_controller import UnifiedUploadController as UploadController
-from .utils.unicode_handler import decode_upload_content
+from unicode_toolkit import decode_upload_content
 from core.unicode import safe_encode_text
 from .ai import AISuggestionService, analyze_device_name_with_ai
 from .helpers import save_ai_training_data
